Creating Complaint style file
