+++ src/Styles/ComplaintStyle.js
@@ -1,0 +1,40 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    complaintScreen: {
+      flex: 1,
+    },
+  
+    textBox: {
+      marginTop: 20,
+      paddingLeft: 2,
+      justifyContent: 'flex-start',
+    },
+  
+    field: {
+      backgroundColor: '#FAFAFA',
+      paddingVertical: 4,
+      borderWidth: 1,
+      borderRadius: 7,
+      borderColor: 'gray',
+      marginHorizontal: 15,
+      marginBottom: 15,
+      marginTop: 15,
+      justifyContent: 'flex-start',
+      paddingLeft: 2,
+      paddingRight: 4,
+    },
+  
+    information: {
+      fontSize: 15,
+    },
+  
+    title: {
+      fontSize: 16,
+      fontWeight: 'bold',
+      textAlign: 'center',
+    },
+  
+  });
+
+  export default styles;
