+++ src/screens/complaintScreen.js
@@ -4,43 +4,7 @@ import Hyperlink from 'react-native-hyperlink';
 import Header from '../components/Header';
 import * as constant from '../constants/publicAgencyInformations';
 import { backHandlerPopToMain } from '../NavigationFunctions';
-
-const styles = StyleSheet.create({
-  complaintScreen: {
-    flex: 1,
-  },
-
-  textBox: {
-    marginTop: 20,
-    paddingLeft: 2,
-    justifyContent: 'flex-start',
-  },
-
-  field: {
-    backgroundColor: '#FAFAFA',
-    paddingVertical: 4,
-    borderWidth: 1,
-    borderRadius: 7,
-    borderColor: 'gray',
-    marginHorizontal: 15,
-    marginBottom: 15,
-    marginTop: 15,
-    justifyContent: 'flex-start',
-    paddingLeft: 2,
-    paddingRight: 4,
-  },
-
-  information: {
-    fontSize: 15,
-  },
-
-  title: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-
-});
+import styles from '../Styles/RegisterStyles';
 
 class complaintScreen extends React.PureComponent {
   componentWillMount() {
